commands/stop.py: refactored boilerplate, added flag for basic stop, added feature to delete dangling images

+++ vm_lifecycle/commands/stop.py
@@ -1,45 +1,28 @@
 import click
 
-from vm_lifecycle.config_manager import ConfigManager
-from vm_lifecycle.compute_manager import GCPComputeManager
-from vm_lifecycle.utils import poll_with_spinner
+# from vm_lifecycle.config_manager import ConfigManager
+# from vm_lifecycle.compute_manager import GCPComputeManager
+from vm_lifecycle.utils import poll_with_spinner, init_gcp_context
 
 
 @click.command(name="stop")
 @click.option("-k", "--keep", is_flag=True, help="Do not delete the instance.")
-def stop_vm_instance(keep):
-    # Load Profile
-    config_manager = ConfigManager()
-
-    profile_check = config_manager.pre_run_profile_check()
-    if not profile_check:
-        click.echo("Error with active profile. Ensure a profile has been created.")
+@click.option(
+    "-b",
+    "--basic",
+    is_flag=True,
+    help="Only shut down the VM instance. No images will will be created, no instances will be destroyed",
+)
+def stop_vm_instance(keep, basic):
+    config_manager, compute_manager, active_zone = init_gcp_context()
+    if not config_manager:
         return
-
-    active_zone = config_manager.active_profile["zone"]
-
-    compute_manager = GCPComputeManager(
-        config_manager.active_profile["project_id"],
-        active_zone,
-    )
-
-    # GCP API Check
-    if not config_manager.active_profile["api_cache"]:
-        apis = compute_manager.check_required_apis()
-        if apis["missing"]:
-            click.echo("❗ The following GCP API's have not been enabled:")
-            for api in apis["missing"]:
-                click.echo(f"\t{api}")
-            return
-        config_manager.config[config_manager.active]["api_cache"] = True
-        config_manager.save_config()
 
     # Check if instance exists
     existing_instances = compute_manager.list_instances()
 
     instance_exists = False
     instance_running = False
-
     if existing_instances:
         for instance in existing_instances:
             if (
@@ -71,13 +54,13 @@ def stop_vm_instance(keep):
             f"✅ Instance: '{config_manager.active_profile["instance_name"]}' stopped"
         )
 
-        # TODO: add spinner
         poll_with_spinner(
             compute_manager=compute_manager,
             op_name=op["name"],
             text=spinner_text,
             done_text=done_text,
             scope="zone",
+            zone=active_zone,
         )
 
     elif not instance_running and not instance_exists:
@@ -90,37 +73,19 @@ def stop_vm_instance(keep):
         # Unreachable
         pass
 
-    # Create an image from the stopped instance
-    op = compute_manager.create_image_from_instance(
-        instance_name=config_manager.active_profile["instance_name"],
-        image_name=config_manager.active_profile["image_base_name"],
-        family=config_manager.active_profile["image_base_name"],
-        zone=active_zone,
-    )
+    if not basic:
+        # Create an image from the stopped instance
+        op = compute_manager.create_image_from_instance(
+            instance_name=config_manager.active_profile["instance_name"],
+            image_name=config_manager.active_profile["image_base_name"],
+            family=config_manager.active_profile["image_base_name"],
+            zone=active_zone,
+        )
 
-    spinner_text = f"Creating image from instance: '{config_manager.active_profile['instance_name']}'"
-    done_text = "✅ Image created!"
-    poll_with_spinner(
-        compute_manager=compute_manager,
-        op_name=op["name"],
-        text=spinner_text,
-        done_text=done_text,
-        scope="global",
-    )
+        spinner_text = f"Creating image from instance: '{config_manager.active_profile['instance_name']}'"
 
-    # Delete dangling images
-    dangling_images = compute_manager.get_dangling_images(
-        family=config_manager.active_profile["image_base_name"]
-    )
-    click.echo(
-        f"🗑️  Destroying {len(dangling_images)} dangling image{'s' if len(dangling_images) > 1 else ''}:"
-    )
-    for image in dangling_images:
-        op = compute_manager.delete_image(image)
-
-        spinner_text = f"Destroying image: '{image}'"
-        done_text = f"🗑️  Image: '{image}' destroyed"
-
+        image_name = op["targetLink"].split("/")[-1]
+        done_text = f"✅ Image: '{image_name}' created from instance: '{config_manager.active_profile['instance_name']}'"
         poll_with_spinner(
             compute_manager=compute_manager,
             op_name=op["name"],
@@ -129,8 +94,30 @@ def stop_vm_instance(keep):
             scope="global",
         )
 
+        # Delete dangling images
+        dangling_images = compute_manager.get_dangling_images(
+            family=config_manager.active_profile["image_base_name"]
+        )
+        if dangling_images:
+            click.echo(
+                f"🗑️  Destroying {len(dangling_images)} dangling image{'s' if len(dangling_images) > 1 else ''}:"
+            )
+            for image in dangling_images:
+                op = compute_manager.delete_image(image)
+
+                spinner_text = f"Destroying image: '{image}'"
+                done_text = f"🗑️  Image: '{image}' destroyed"
+
+                poll_with_spinner(
+                    compute_manager=compute_manager,
+                    op_name=op["name"],
+                    text=spinner_text,
+                    done_text=done_text,
+                    scope="global",
+                )
+
     # Delete Compute Engine Instance
-    if not keep:
+    if not keep and not basic:
         op = compute_manager.delete_instance(
             instance_name=config_manager.active_profile["instance_name"],
             zone=active_zone,
@@ -147,4 +134,5 @@ def stop_vm_instance(keep):
             text=spinner_text,
             done_text=done_text,
             scope="zone",
+            zone=active_zone,
         )
